aquila_defconfig: Disable SCHED_WALT
WALT maintains metrics for tasks and runqueues to guide task
placement and frequency requirements, but it doesn't really
work great: this one causes schedutil to always go to a very very
high frequency for reasons, which leads to truly unecessary heat
and battery power waste.

The battery power waste is also due to the fact that WALT seems
to be always putting any medium-intensive task on the fastest CPU,
which isn't a good decision, since the little or big (not big plus) cores
are in 99% of the times fast enough to finish the job very fast and
more efficiently than the big plus / perf plus core(s), outputting less
heat and wasting less battery power.

Disabling this means that we're going PELT: this is actually working
fine now, as only very very intensive tasks will be migrated to the
very very fast and very very "hot" cores.

Signed-off-by: Hachiman270 <hachiman270@gmail.com>